+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/plugin/jei/wrapper/JEIRecipeWrapperCrudeDryingRack.java
@@ -3,37 +3,31 @@ package com.codetaylor.mc.pyrotech.modules.tech.basic.plugin.jei.wrapper;
 import com.codetaylor.mc.pyrotech.library.spi.plugin.jei.JEIRecipeWrapperTimed;
 import com.codetaylor.mc.pyrotech.library.spi.recipe.IRecipeTimed;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasicConfig;
-import com.codetaylor.mc.pyrotech.modules.tech.basic.recipe.SoakingPotRecipe;
+import com.codetaylor.mc.pyrotech.modules.tech.basic.recipe.spi.DryingRackRecipeBase;
 import mezz.jei.api.ingredients.IIngredients;
 import mezz.jei.api.ingredients.VanillaTypes;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.fluids.FluidStack;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
-public class JEIRecipeWrapperSoakingPot
+public class JEIRecipeWrapperCrudeDryingRack
     extends JEIRecipeWrapperTimed {
 
   private final ResourceLocation registryName;
   private final List<List<ItemStack>> inputs;
-  private final FluidStack inputFluid;
   private final ItemStack output;
 
-  public JEIRecipeWrapperSoakingPot(SoakingPotRecipe recipe) {
+  public JEIRecipeWrapperCrudeDryingRack(DryingRackRecipeBase recipe) {
 
     super(recipe);
 
     this.registryName = recipe.getRegistryName();
-
-    this.inputs = new ArrayList<>(2);
-    this.inputs.add(Arrays.asList(recipe.getInputItem().getMatchingStacks()));
-    this.inputFluid = recipe.getInputFluid();
-
+    this.inputs = Collections.singletonList(Arrays.asList(recipe.getInput().getMatchingStacks()));
     this.output = recipe.getOutput();
   }
 
@@ -41,14 +35,13 @@ public class JEIRecipeWrapperSoakingPot
   public void getIngredients(@Nonnull IIngredients ingredients) {
 
     ingredients.setInputLists(VanillaTypes.ITEM, this.inputs);
-    ingredients.setInput(VanillaTypes.FLUID, this.inputFluid);
     ingredients.setOutput(VanillaTypes.ITEM, this.output);
   }
 
   @Override
   protected int getTimeDisplayY() {
 
-    return super.getTimeDisplayY() - 2;
+    return super.getTimeDisplayY() - 14;
   }
 
   @Nullable
@@ -61,6 +54,6 @@ public class JEIRecipeWrapperSoakingPot
   @Override
   protected int getRecipeTime(IRecipeTimed recipe) {
 
-    return (int) (super.getRecipeTime(recipe) * ModuleTechBasicConfig.SOAKING_POT.BASE_RECIPE_DURATION_MODIFIER);
+    return (int) (super.getRecipeTime(recipe) * ModuleTechBasicConfig.CRUDE_DRYING_RACK.BASE_RECIPE_DURATION_MODIFIER);
   }
 }