+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/refractory/plugin/jei/wrapper/JEIRecipeWrapperPitBurn.java
@@ -1,9 +1,11 @@
 package com.codetaylor.mc.pyrotech.modules.tech.refractory.plugin.jei.wrapper;
 
 import com.codetaylor.mc.pyrotech.library.spi.plugin.jei.JEIRecipeWrapperTimed;
+import com.codetaylor.mc.pyrotech.library.spi.recipe.IRecipeTimed;
 import com.codetaylor.mc.pyrotech.library.util.BlockMetaMatcher;
 import com.codetaylor.mc.pyrotech.library.util.Util;
 import com.codetaylor.mc.pyrotech.modules.tech.refractory.ModuleTechRefractory;
+import com.codetaylor.mc.pyrotech.modules.tech.refractory.ModuleTechRefractoryConfig;
 import com.codetaylor.mc.pyrotech.modules.tech.refractory.recipe.PitBurnRecipe;
 import mezz.jei.api.ingredients.IIngredients;
 import mezz.jei.api.ingredients.VanillaTypes;
@@ -95,5 +97,11 @@ public class JEIRecipeWrapperPitBurn
   public ResourceLocation getRegistryName() {
 
     return this.registryName;
+  }
+
+  @Override
+  protected int getRecipeTime(IRecipeTimed recipe) {
+
+    return (int) (super.getRecipeTime(recipe) * ModuleTechRefractoryConfig.REFRACTORY.BASE_RECIPE_DURATION_MODIFIER);
   }
 }